backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
langchain
openai
faiss-cpu